pyproject.toml
[project]
name = "mcp-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]
